Update function signatures

+++ containers.h
@@ -144,22 +144,21 @@ unordered_multiset unordered_multiset_destroy(unordered_multiset me);
 typedef struct internal_multimap *multimap;
 
 /* Starting */
-multimap multimap_init(size_t key_size,
-                       size_t value_size,
+multimap multimap_init(size_t key_size, size_t value_size,
                        int (*key_comparator)(const void *const one,
                                              const void *const two),
                        int (*value_comparator)(const void *const one,
                                                const void *const two));
 
 /* Capacity */
-int multimap_size(multimap me);
+size_t multimap_size(multimap me);
 int multimap_is_empty(multimap me);
 
 /* Accessing */
 int multimap_put(multimap me, void *key, void *value);
 void multimap_get_start(multimap me, void *key);
 int multimap_get_next(void *value, multimap me);
-int multimap_count(multimap me, void *key);
+size_t multimap_count(multimap me, void *key);
 int multimap_contains(multimap me, void *key);
 int multimap_remove(multimap me, void *key, void *value);
 int multimap_remove_all(multimap me, void *key);
